add struct to text field and uiview
