+++ Bankodemia/Models/Constants/ConstantsColor.swift
@@ -10,8 +10,13 @@ import UIKit
 
 struct ConstantsCGColor{
     static let greenBlue : CGColor = CGColor(red: 0, green: 204/255, blue: 204/255, alpha: 0.9)
+    
+    static let kodemiaGrey : CGColor = CGColor(red: 127/255, green: 127/255, blue: 127/255, alpha: 0.9)
 }
 
 struct ConstantsUIColor{
     static let greenBlue : UIColor = UIColor(red: 0, green: 204/255, blue: 204/255, alpha: 0.9)
+    
+    static let kodemiaGrey : UIColor = UIColor(red: 127/255, green: 127/255, blue: 127/255, alpha: 0.9)
+
 }
