+++ Bankodemia/Modules/Login/LoginViewController.swift
@@ -11,6 +11,7 @@ class LoginViewController: UIViewController {
 
     override func viewDidLoad() {
         super.viewDidLoad()
+        view.backgroundColor = .red
         // Do any additional setup after loading the view.
     }
